# Improve perf for positional mask