- improve perfs for the positional mask
- positional mask still correct if ids are not monotonic starting from 0

Tests : 
```
df = pd.DataFrame(index=np.arange(5000000))

test_ids = np.random.randint(len(df), size=50)
%timeit _positional_mask_old(df, test_ids)
115 ms ± 910 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
%timeit _positional_mask(df, test_ids)
95.2 µs ± 417 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

test_ids = np.random.randint(len(df), size=50000)
%timeit _positional_mask_old(df, test_ids)
141 ms ± 3.89 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
%timeit _positional_mask(df, test_ids)
2.32 ms ± 87.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

test_ids = np.random.randint(len(df), size=3000000)
%timeit _positional_mask_old(df, test_ids)
1.05 s ± 15 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
%timeit _positional_mask(df, test_ids)
111 ms ± 2.33 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

df = pd.DataFrame(index=np.arange(50000000))
test_ids = np.random.randint(len(df), size=3000000)
%timeit _positional_mask_old(df, test_ids)
2.45 s ± 24 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit _positional_mask(df, test_ids)
112 ms ± 2.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```